# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Church Management System API
  description: |
    API complète pour la gestion d'église moderne et sécurisée.
    
    ## 🔒 Système CSRF Automatique
    - **Aucune configuration manuelle requise** pour les tokens CSRF
    - **Génération automatique** à la connexion
    - **Envoi automatique** sur toutes les requêtes protégées
    - **Validation automatique** côté serveur
    - **Expiration automatique** avec déconnexion intelligente
    
    ## 🚀 Fonctionnalités
    - Gestion complète des utilisateurs et églises
    - Système de réseaux et groupes
    - Statistiques en temps réel
    - Gestion des services et cultes
    - Système de permissions avancé
    
    ## 🗄️ Base de données
    - **PostgreSQL** avec Prisma ORM
    - **Migration automatique** depuis MongoDB
    - **Relations complexes** gérées nativement
  version: 2.0.0
  contact:
    name: Church Management System
    email: contact@church-system.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: ${APP_BACKEND_URL:-http://localhost:5001}/api
    description: Serveur de développement
  - url: ${APP_BACKEND_URL:-http://localhost:5001}/api
    description: Serveur de production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenu lors de la connexion.
        Inclut automatiquement le token CSRF dans le header X-CSRF-Token.
    
    CSRFProtection:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Token CSRF généré automatiquement à la connexion.
        **Aucune action manuelle requise** - géré automatiquement par le système.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Message d'erreur lisible
        code:
          type: string
          description: Code d'erreur spécifique
        error:
          type: string
          description: Détails techniques (optionnel)
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Message de succès
        data:
          description: Données de la réponse
...
  responses:
    UnauthorizedError:
      description: Token d'authentification manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Token expiré ou invalide"
            code: "JWT_INVALID"
    ForbiddenError:
      description: Permissions insuffisantes pour accéder à la ressource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Accès refusé : permissions insuffisantes"
            code: "INSUFFICIENT_PERMISSIONS"
...
security:
  - BearerAuth: []
  - CSRFProtection: []
paths:
...
  /admin/clear-cache:
    post:
      tags:
        - Admin
      summary: Nettoyer le cache
      description: |
        Nettoie le cache serveur (Admin uniquement).
        
        **Permissions requises :**
        - **ADMIN** ou **SUPER_ADMIN**
      security:
        - BearerAuth: []
        - CSRFProtection: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pattern:
                  type: string
                  description: Pattern pour nettoyer une partie spécifique du cache
                  example: "stats"
      responses:
        '200':
          description: Cache nettoyé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /admin/cache-status:
    get:
      tags:
        - Admin
      summary: Voir l'état du cache
      description: |
        Affiche l'état actuel du cache serveur (Admin uniquement).
        
        **Permissions requises :**
        - **ADMIN** ou **SUPER_ADMIN**
      security:
        - BearerAuth: []
        - CSRFProtection: []
      responses:
        '200':
          description: État du cache récupéré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...